Backend/requirements.txt
fastapi
uvicorn
rembg
numpy
opencv-python-headless
pillow
